+++ helpers/src/edhoc_helpers.c
@@ -23,7 +23,7 @@ LOG_MODULE_DECLARE(libedhoc, CONFIG_LIBEDHOC_LOG_LEVEL);
 #include "edhoc.h"
 #include "edhoc_helpers.h"
 #include "edhoc_common.h"
-#include "edhoc_log.h"
+#include "edhoc_backend_log.h"
 
 #ifdef __clang__
 #pragma clang diagnostic push